pyproject.toml
[project]
name = "beeminder-wc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
